+++ internal/pkg/instrumentation/probe/probe.go
@@ -20,6 +20,7 @@ import (
 
 	"go.opentelemetry.io/auto/internal/pkg/inject"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/bpffs"
+	"go.opentelemetry.io/auto/internal/pkg/instrumentation/probe/sampling"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/utils"
 	"go.opentelemetry.io/auto/internal/pkg/process"
 	"go.opentelemetry.io/auto/internal/pkg/structfield"
@@ -32,8 +33,11 @@ type Probe interface {
 	// the information about the package the Probe instruments.
 	Manifest() Manifest
 
-	// Load loads all instrumentation offsets.
-	Load(*link.Executable, *process.TargetDetails) error
+	// Load loads all the eBPF programs and maps required by the Probe.
+	// It also attaches the eBPF programs to the target process.
+	// TODO: currently passing Sampler as an initial configuration - this will be
+	// updated to a more generic configuration in the future.
+	Load(*link.Executable, *process.TargetDetails, *sampling.Config) error
 
 	// Run runs the events processing loop.
 	Run(eventsChan chan<- *Event)
@@ -92,14 +96,18 @@ func (i *Base[BPFObj, BPFEvent]) Manifest() Manifest {
 	return NewManifest(i.ID, structfields, symbols)
 }
 
+func (i *Base[BPFObj, BPFEvent]) Spec() (*ebpf.CollectionSpec, error) {
+	return i.SpecFn()
+}
+
 // Load loads all instrumentation offsets.
-func (i *Base[BPFObj, BPFEvent]) Load(exec *link.Executable, td *process.TargetDetails) error {
+func (i *Base[BPFObj, BPFEvent]) Load(exec *link.Executable, td *process.TargetDetails, sampler *sampling.Config) error {
 	spec, err := i.SpecFn()
 	if err != nil {
 		return err
 	}
 
-	err = i.injectConsts(td, spec)
+	err = i.InjectConsts(td, spec)
 	if err != nil {
 		return err
 	}
@@ -118,12 +126,17 @@ func (i *Base[BPFObj, BPFEvent]) Load(exec *link.Executable, td *process.TargetD
 	if err != nil {
 		return err
 	}
+
+	// TODO: Initialize sampling manager based on the sampling configuration and the eBPF collection.
+	// The manager will be responsible for writing to eBPF maps - configuring the sampling.
+	// In addition the sampling manager will be responsible for handling updates for the configuration.
+
 	i.closers = append(i.closers, i.reader)
 
 	return nil
 }
 
-func (i *Base[BPFObj, BPFEvent]) injectConsts(td *process.TargetDetails, spec *ebpf.CollectionSpec) error {
+func (i *Base[BPFObj, BPFEvent]) InjectConsts(td *process.TargetDetails, spec *ebpf.CollectionSpec) error {
 	opts, err := consts(i.Consts).injectOpts(td)
 	if err != nil {
 		return err